updated markup for Booking history
